move 공통응답클래스를 common패키지로 옮김
